Tweak position of auto popup button and menu
Make the auto popup button appear slightly under the mouse (except when
near the bottom of the browser window), which makes the button easier
to locate and move to.

Also, modify the position of the resulting popup menu so that, when
using the "No selection" styling option, the name of the first engine
appears under the mouse. This makes the first engine easy to select
while not covering up any icons with the mouse.

Since this new menu positioning does not make very much sense with
styling options other than the “No selection” option, this change
assumes that the user will be using the “No selection” option.