Allow custom repositioning of popup at edge
Add style config options for repositioning the menu or button whenever
either is outside of the browser window. This can prevent the menu from
covering the selection, or the auto button from appearing directly
under the mouse, whenever the selection is next to one of the window’s
edges. See Issue #15 for discussion.