Possible fix for freezing new tab
There has been some reports that a search opened in a new tab
sometimes freezed for a while after the new tab has been opened.

This olly seems to happen when using the open search in new tab
option, without using the open search in background tab.

This change seems to maybe fix the issue.