Reinject the popup element if it gets removed
Some pages may replace the page content after the initial page load,
this can sometimes remove the selection search html elements. This
adds a check to try to reinject the popup a few times if it gets
removed.

This is just using a timer to check for a few seconds after the initial load.
If the popup is removed after several seconds it will not be handled
by this fix.